+++ server/src/http/routes/etablissement.controller.ts
@@ -11,6 +11,7 @@ import { tryCatch } from "../middlewares/tryCatchMiddleware.js"
 import * as eligibleTrainingsForAppointmentService from "../../services/eligibleTrainingsForAppointment.service.js"
 import * as appointmentService from "../../services/appointment.service.js"
 import mailer from "../../services/mailer.service.js"
+import { Etablissement } from "../../common/model/index.js"
 
 const optOutUnsubscribeSchema = Joi.object({
   opt_out_question: Joi.string().optional(),
@@ -23,7 +24,7 @@ const patchEtablissementIdAppointmentIdReadAppointSchema = Joi.object({
 /**
  * @description Etablissement Router.
  */
-export default ({ etablissements }) => {
+export default () => {
   const router = express.Router()
 
   /**
@@ -32,7 +33,7 @@ export default ({ etablissements }) => {
   router.get(
     "/:id",
     tryCatch(async (req, res) => {
-      const etablissement = await etablissements.findById(req.params.id)
+      const etablissement = await Etablissement.findById(req.params.id)
 
       if (!etablissement) {
         return res.sendStatus(404)
@@ -48,7 +49,7 @@ export default ({ etablissements }) => {
   router.post(
     "/:id/premium/affelnet/accept",
     tryCatch(async (req, res) => {
-      const etablissement = await etablissements.findById(req.params.id)
+      const etablissement = await Etablissement.findById(req.params.id)
 
       if (!etablissement) {
         throw Boom.badRequest("Etablissement not found.")
@@ -84,7 +85,7 @@ export default ({ etablissements }) => {
         eligibleTrainingsForAppointmentService.find({
           etablissement_formateur_siret: etablissement.formateur_siret,
         }),
-        etablissements.findOneAndUpdate(
+        Etablissement.findOneAndUpdate(
           { _id: etablissement._id },
           {
             $push: {
@@ -138,7 +139,7 @@ export default ({ etablissements }) => {
       )
 
       const [resultAffelnet] = await Promise.all([
-        etablissements.findById(req.params.id),
+        Etablissement.findById(req.params.id),
         ...eligibleTrainingsForAppointmentsAffelnetFound.map((eligibleTrainingsForAppointment) =>
           eligibleTrainingsForAppointmentService.update(
             { _id: eligibleTrainingsForAppointment._id, lieu_formation_email: { $nin: [null, ""] } },
@@ -159,7 +160,7 @@ export default ({ etablissements }) => {
   router.post(
     "/:id/premium/accept",
     tryCatch(async (req, res) => {
-      const etablissement = await etablissements.findById(req.params.id)
+      const etablissement = await Etablissement.findById(req.params.id)
 
       if (!etablissement) {
         throw Boom.badRequest("Etablissement not found.")
@@ -198,7 +199,7 @@ export default ({ etablissements }) => {
             $ne: null,
           },
         }),
-        etablissements.findOneAndUpdate(
+        Etablissement.findOneAndUpdate(
           { _id: etablissement._id },
           {
             $push: {
@@ -252,7 +253,7 @@ export default ({ etablissements }) => {
       )
 
       const [result] = await Promise.all([
-        etablissements.findById(req.params.id),
+        Etablissement.findById(req.params.id),
         ...eligibleTrainingsForAppointmentsParcoursupFound.map((eligibleTrainingsForAppointment) =>
           eligibleTrainingsForAppointmentService.update(
             { _id: eligibleTrainingsForAppointment._id, lieu_formation_email: { $nin: [null, ""] } },
@@ -273,7 +274,7 @@ export default ({ etablissements }) => {
   router.post(
     "/:id/premium/affelnet/refuse",
     tryCatch(async (req, res) => {
-      const etablissement = await etablissements.findById(req.params.id)
+      const etablissement = await Etablissement.findById(req.params.id)
 
       if (!etablissement) {
         throw Boom.badRequest("Etablissement not found.")
@@ -310,7 +311,7 @@ export default ({ etablissements }) => {
         },
       })
 
-      await etablissements.findOneAndUpdate(
+      await Etablissement.findOneAndUpdate(
         { _id: etablissement._id },
         {
           $push: {
@@ -325,7 +326,7 @@ export default ({ etablissements }) => {
         }
       )
 
-      const etablissementAffelnetUpdated = await etablissements.findById(req.params.id)
+      const etablissementAffelnetUpdated = await Etablissement.findById(req.params.id)
 
       return res.send(etablissementAffelnetUpdated)
     })
@@ -337,7 +338,7 @@ export default ({ etablissements }) => {
   router.post(
     "/:id/premium/refuse",
     tryCatch(async (req, res) => {
-      const etablissement = await etablissements.findById(req.params.id)
+      const etablissement = await Etablissement.findById(req.params.id)
 
       if (!etablissement) {
         throw Boom.badRequest("Etablissement not found.")
@@ -374,7 +375,7 @@ export default ({ etablissements }) => {
         },
       })
 
-      await etablissements.findOneAndUpdate(
+      await Etablissement.findOneAndUpdate(
         { _id: etablissement._id },
         {
           $push: {
@@ -389,7 +390,7 @@ export default ({ etablissements }) => {
         }
       )
 
-      const etablissementParcoursupUpdated = await etablissements.findById(req.params.id)
+      const etablissementParcoursupUpdated = await Etablissement.findById(req.params.id)
 
       return res.send(etablissementParcoursupUpdated)
     })
@@ -408,7 +409,7 @@ export default ({ etablissements }) => {
       const { id, appointmentId } = params
 
       // eslint-disable-next-line prefer-const
-      let [etablissement, appointment] = await Promise.all([etablissements.findById(id), appointmentService.findById(appointmentId)])
+      let [etablissement, appointment] = await Promise.all([Etablissement.findById(id), appointmentService.findById(appointmentId)])
 
       if (!etablissement) {
         throw Boom.badRequest("Etablissement not found.")
@@ -437,7 +438,7 @@ export default ({ etablissements }) => {
     tryCatch(async (req, res) => {
       const { opt_out_question } = await optOutUnsubscribeSchema.validateAsync(req.body, { abortEarly: false })
 
-      let etablissement = await etablissements.findById(req.params.id)
+      let etablissement = await Etablissement.findById(req.params.id)
 
       if (!etablissement) {
         return res.sendStatus(404)
@@ -448,7 +449,7 @@ export default ({ etablissements }) => {
       }
 
       if (opt_out_question) {
-        etablissement = await etablissements.findById(req.params.id)
+        etablissement = await Etablissement.findById(req.params.id)
 
         await mailer.sendEmail({
           to: config.publicEmail,
@@ -489,7 +490,7 @@ export default ({ etablissements }) => {
         )
       }
 
-      await etablissements.findByIdAndUpdate(req.params.id, {
+      await Etablissement.findByIdAndUpdate(req.params.id, {
         optout_refusal_date: dayjs().toDate(),
       })
 
@@ -515,7 +516,7 @@ export default ({ etablissements }) => {
         },
       })
 
-      await etablissements.findOneAndUpdate(
+      await Etablissement.findOneAndUpdate(
         { _id: etablissement._id },
         {
           $push: {
@@ -529,7 +530,7 @@ export default ({ etablissements }) => {
         }
       )
 
-      etablissement = await etablissements.findById(req.params.id)
+      etablissement = await Etablissement.findById(req.params.id)
 
       return res.send(etablissement)
     })