+++ ui/components/SearchForTrainingsAndJobs/components/SearchForm.js
@@ -76,7 +76,7 @@ const SearchForm = (props) => {
             <Box p="0">
               <Text as="h1" fontSize={["26px", "29px"]} fontWeight={700}>
                 <Text as="span" display={{ base: "block", md: "inline" }}>
-                  Trouvez emploi et formation{" "}
+                  Emplois et formations{" "}
                 </Text>
                 <Text as="span" color="info" display={{ base: "block", md: "inline" }}>
                   en alternance